Partially watched blue color replaced with grey, selector backgrounds added and selection persistent background removed
