+++ VideoLocker/res/drawable/course_list_item_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@color/edx_grayscale_neutral_light" android:state_pressed="true"/>
+    <item android:drawable="@color/white"/>
+</selector>
