+++ VideoLocker/res/layout/row_course_dashboard_list.xml
@@ -4,7 +4,7 @@
     android:id="@+id/chapter_row_layout"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:background="@drawable/rectangle_with_bottom_border"
+    android:background="@drawable/rectangle_with_bottom_border_selector"
     android:gravity="center_vertical"
     android:layout_marginStart="16dp"
     android:layout_marginLeft="16dp"
