+++ VideoLocker/src/main/java/org/edx/mobile/view/adapters/CourseOutlineAdapter.java
@@ -2,16 +2,21 @@ package org.edx.mobile.view.adapters;
 
 import android.content.Context;
 import android.text.TextUtils;
+import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.BaseAdapter;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import org.edx.mobile.R;
 import org.edx.mobile.base.MainApplication;
 import org.edx.mobile.event.DownloadEvent;
+import org.edx.mobile.logger.Logger;
 import org.edx.mobile.model.course.BlockPath;
 import org.edx.mobile.model.course.BlockType;
 import org.edx.mobile.model.course.CourseComponent;
+import org.edx.mobile.model.course.HasDownloadEntry;
 import org.edx.mobile.model.course.IBlock;
 import org.edx.mobile.model.course.VideoBlockModel;
 import org.edx.mobile.model.db.DownloadEntry;
@@ -20,9 +25,12 @@ import org.edx.mobile.module.db.DataCallback;
 import org.edx.mobile.module.db.IDatabase;
 import org.edx.mobile.module.prefs.PrefManager;
 import org.edx.mobile.module.storage.IStorage;
+import org.edx.mobile.third_party.iconify.IconView;
 import org.edx.mobile.third_party.iconify.Iconify;
 import org.edx.mobile.util.AppConstants;
+import org.edx.mobile.view.custom.ETextView;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import de.greenrobot.event.EventBus;
@@ -30,13 +38,99 @@ import de.greenrobot.event.EventBus;
 /**
  * Used for pinned behavior.
  */
-public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
+public class CourseOutlineAdapter extends BaseAdapter{
+
+    private final Logger logger = new Logger(getClass().getName());
+
+    public interface DownloadListener {
+        void download(List<HasDownloadEntry> models);
+        void download(DownloadEntry videoData);
+    }
+
+    private CourseComponent rootComponent;
+    private LayoutInflater mInflater;
+    private List<SectionRow> mData;
+
+    private IDatabase dbStore;
+    private IStorage storage;
+    private DownloadListener mDownloadListener;
+    private Context context;
 
     private boolean currentVideoMode;
     private int numOfTotalUnits;
 
-    public CourseOutlineAdapter(Context context, IDatabase dbStore, IStorage storage) {
-        super(context, dbStore, storage);
+    public CourseOutlineAdapter(Context context, IDatabase dbStore, IStorage storage, DownloadListener listener) {
+        this.context = context;
+        this.dbStore = dbStore;
+        this.storage = storage;
+        this.mDownloadListener = listener;
+        mInflater = (LayoutInflater)context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        mData = new ArrayList();
+    }
+
+    @Override public int getItemViewType(int position) {
+        return getItem(position).type;
+    }
+
+    @Override
+    public int getViewTypeCount() {
+        return 2;
+    }
+
+    @Override
+    public int getCount() {
+        return mData.size();
+    }
+
+    @Override
+    public SectionRow getItem(int position) {
+        if ( position < 0 || position >= mData.size() )
+            return null;
+        return mData.get(position);
+    }
+
+    @Override
+    public long getItemId(int position) {
+        return position;
+    }
+
+    @Override
+    public boolean isEnabled(int position) {
+        return getItemViewType(position) == SectionRow.ITEM;
+    }
+
+    @Override
+    public boolean areAllItemsEnabled() {
+        return false;
+    }
+
+    @Override
+    public final View getView(int position, View convertView, ViewGroup parent) {
+
+        int type = getItemViewType(position);
+
+        if (convertView == null) {
+            switch (type) {
+                case SectionRow.ITEM: {
+                    convertView = mInflater.inflate(R.layout.row_course_outline_list, parent, false);
+                    // apply a tag to this list row
+                    ViewHolder tag = getTag(convertView);
+                    convertView.setTag(tag);
+                    break;
+                }
+                default: {//SectionRow.SECTION:
+                    convertView = mInflater.inflate(R.layout.row_section_header, parent, false);
+                    break;
+                }
+            }
+        }
+
+        switch (type) {
+            case SectionRow.ITEM:
+                return  getRowView(position, convertView, parent);
+            default : //SectionRow.SECTION:
+                return  getHeaderView(position, convertView, parent);
+        }
     }
 
     /**
@@ -78,16 +172,15 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
         notifyDataSetChanged();
     }
 
-    @Override
+    public void reloadData(){
+        if (  this.rootComponent != null )
+            setData(this.rootComponent);
+    }
+
     public View getRowView(int position, View convertView, ViewGroup parent) {
         final SectionRow row = this.getItem(position);
         final SectionRow nextRow = this.getItem(position+1);
         final CourseComponent component = row.component;
-        if (selectedRow != null && selectedRow.component.equals( row.component )){
-            convertView.setBackgroundResource(R.color.edx_brand_primary_x_light);
-        } else {
-            convertView.setBackgroundResource(R.color.white);
-        }
         final ViewHolder viewHolder = (ViewHolder)convertView.getTag();
 
         if ( nextRow == null ){
@@ -103,8 +196,6 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
             }
         }
 
-
-
         viewHolder.rowType.setVisibility(View.GONE);
         viewHolder.rowSubtitleIcon.setVisibility(View.GONE);
         viewHolder.rowSubtitle.setVisibility(View.GONE);
@@ -119,7 +210,6 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
     }
 
     private  View getRowViewForLeaf(int position, View convertView, ViewGroup parent, final SectionRow row) {
-
         final ViewHolder viewHolder = (ViewHolder)convertView.getTag();
 
         CourseComponent unit =  row.component;
@@ -153,25 +243,20 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
         } else {
             viewHolder.rowTitle.setText(unit.getDisplayName());
         }
-        convertView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                rowClicked(row);
-            }
-        });
         return convertView;
     }
 
-    private void checkAccessStatus(final ViewHolder viewHolder, final CourseComponent unit){
-        dbStore.isUnitAccessed(new DataCallback<Boolean>(true){
+    private void checkAccessStatus(final ViewHolder viewHolder, final CourseComponent unit) {
+        dbStore.isUnitAccessed(new DataCallback<Boolean>(true) {
             @Override
             public void onResult(Boolean accessed) {
-                if( accessed ) {
+                if (accessed) {
                     viewHolder.rowType.setIconColor(context.getResources().getColor(R.color.edx_grayscale_neutral_base));
                 } else {
                     viewHolder.rowType.setIconColor(context.getResources().getColor(R.color.edx_brand_primary_base));
                 }
             }
+
             @Override
             public void onFail(Exception ex) {
                 logger.error(ex);
@@ -212,7 +297,7 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
                     if(ws == null || ws == DownloadEntry.WatchedState.UNWATCHED) {
                         viewHolder.rowType.setIconColor(context.getResources().getColor(R.color.edx_brand_primary_base));
                     } else if(ws == DownloadEntry.WatchedState.PARTIALLY_WATCHED) {
-                        viewHolder.rowType.setIconColor(context.getResources().getColor(R.color.cyan_2));
+                        viewHolder.rowType.setIconColor(context.getResources().getColor(R.color.edx_grayscale_neutral_base));
                     } else {
                         viewHolder.rowType.setIconColor(context.getResources().getColor(R.color.edx_grayscale_neutral_base));
                     }
@@ -224,56 +309,57 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
             });
 
         dbStore.getDownloadedStateForVideoId(videoData.videoId,
-            new DataCallback<DownloadEntry.DownloadedState>(true) {
-                @Override
-                public void onResult(DownloadEntry.DownloadedState result) {
-                    DownloadEntry.DownloadedState ds = result;
-                    if(ds == null || ds == DownloadEntry.DownloadedState.ONLINE) {
-                        // not yet downloaded
-                        viewHolder.bulkDownload.setVisibility(View.VISIBLE);
-                        viewHolder.numOfVideoAndDownloadArea.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.STARTED));
-                                logger.debug("Download Button Clicked");
-                                //notifyDataSetChanged();
-                                download(videoData);
-                            }
-                        });
-                    } else if(ds == DownloadEntry.DownloadedState.DOWNLOADING) {
-                        // may be download in progress
-                        EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.STARTED));
-                        viewHolder.bulkDownload.setVisibility(View.GONE);
-                        storage.getDownloadProgressByDmid(videoData.dmId, new DataCallback<Integer>(true) {
-                            @Override
-                            public void onResult(Integer result) {
-                                if(result>=0 && result < 100){
+                new DataCallback<DownloadEntry.DownloadedState>(true) {
+                    @Override
+                    public void onResult(DownloadEntry.DownloadedState result) {
+                        DownloadEntry.DownloadedState ds = result;
+                        if (ds == null || ds == DownloadEntry.DownloadedState.ONLINE) {
+                            // not yet downloaded
+                            viewHolder.bulkDownload.setVisibility(View.VISIBLE);
+                            viewHolder.numOfVideoAndDownloadArea.setOnClickListener(new View.OnClickListener() {
+                                @Override
+                                public void onClick(View v) {
                                     EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.STARTED));
-                                } else if ( result == 100 ){
-                                    EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.COMPLETED));
+                                    logger.debug("Download Button Clicked");
+                                    //notifyDataSetChanged();
+                                    mDownloadListener.download(videoData);
                                 }
-                            }
-                            @Override
-                            public void onFail(Exception ex) {
-                                logger.error(ex);
-                                viewHolder.bulkDownload.setVisibility(View.VISIBLE);
-                            }
-                        });
-                    } else if (ds == DownloadEntry.DownloadedState.DOWNLOADED) {
-                        // downloaded
-                        viewHolder.bulkDownload.setVisibility(View.GONE);
+                            });
+                        } else if (ds == DownloadEntry.DownloadedState.DOWNLOADING) {
+                            // may be download in progress
+                            EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.STARTED));
+                            viewHolder.bulkDownload.setVisibility(View.GONE);
+                            storage.getDownloadProgressByDmid(videoData.dmId, new DataCallback<Integer>(true) {
+                                @Override
+                                public void onResult(Integer result) {
+                                    if (result >= 0 && result < 100) {
+                                        EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.STARTED));
+                                    } else if (result == 100) {
+                                        EventBus.getDefault().post(new DownloadEvent(DownloadEvent.DownloadStatus.COMPLETED));
+                                    }
+                                }
+
+                                @Override
+                                public void onFail(Exception ex) {
+                                    logger.error(ex);
+                                    viewHolder.bulkDownload.setVisibility(View.VISIBLE);
+                                }
+                            });
+                        } else if (ds == DownloadEntry.DownloadedState.DOWNLOADED) {
+                            // downloaded
+                            viewHolder.bulkDownload.setVisibility(View.GONE);
+                        }
+
                     }
 
-                }
-                @Override
-                public void onFail(Exception ex) {
-                    logger.error(ex);
-                    viewHolder.bulkDownload.setVisibility(View.VISIBLE);
-                }
-            });
+                    @Override
+                    public void onFail(Exception ex) {
+                        logger.error(ex);
+                        viewHolder.bulkDownload.setVisibility(View.VISIBLE);
+                    }
+                });
 
     }
-
 
     private  View getRowViewForContainer(int position, View convertView, ViewGroup parent, final SectionRow row) {
         final CourseComponent component = row.component;
@@ -285,17 +371,8 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
         String sequentialId =  path.get(2) == null ? "" : path.get(2).getDisplayName();
 
         ViewHolder holder = (ViewHolder)convertView.getTag();
-
         holder.rowTitle.setText(component.getDisplayName());
-        convertView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                rowClicked(row);
-            }
-        });
-
         holder.numOfVideoAndDownloadArea.setVisibility(View.VISIBLE);
-
         if ( component.isGraded() ){
             holder.bulkDownload.setVisibility(View.INVISIBLE);
             holder.rowSubtitlePanel.setVisibility(View.VISIBLE);
@@ -303,7 +380,6 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
             holder.rowSubtitle.setVisibility(View.VISIBLE);
             holder.rowSubtitle.setText(component.getFormat());
         }
-
 
         //support video download for video type
         final int totalCount = component.getBlockCount().videoCount;
@@ -325,7 +401,7 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
                         .setOnClickListener(new View.OnClickListener() {
                             @Override
                             public void onClick(View downloadView) {
-                                download(component.getVideos());
+                                mDownloadListener.download(component.getVideos());
                             }
                         });
                 } else {
@@ -344,9 +420,8 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
             }else{
                 //TODO - any UI update
             }
-
         } else {
-           //TOOD - any UI update?
+           //TODO - any UI update?
         }
 
         return convertView;
@@ -357,7 +432,7 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
         TextView titleView = (TextView)convertView.findViewById(R.id.row_header);
         View separator = convertView.findViewById(R.id.row_separator);
         titleView.setText(row.component.getDisplayName());
-        if ( position == 0 ){
+        if ( position == 0) {
             separator.setVisibility(View.GONE);
         } else {
             separator.setVisibility(View.VISIBLE);
@@ -385,5 +460,66 @@ public abstract  class CourseOutlineAdapter extends CourseBaseAdapter  {
      */
     public boolean hasFilteredUnits(){
         return this.numOfTotalUnits > mData.size();
+    }
+
+    public ViewHolder getTag(View convertView) {
+        ViewHolder holder = new ViewHolder();
+        holder.rowType = (IconView) convertView
+                .findViewById(R.id.row_type);
+        holder.rowTitle = (ETextView) convertView
+                .findViewById(R.id.row_title);
+        holder.rowSubtitle = (ETextView) convertView
+                .findViewById(R.id.row_subtitle);
+        holder.rowSubtitleIcon = (IconView) convertView
+                .findViewById(R.id.row_subtitle_icon);
+        holder.rowSubtitleIcon.setIconColor(context.getResources().getColor(R.color.edx_grayscale_neutral_light));
+        holder.noOfVideos = (TextView) convertView
+                .findViewById(R.id.no_of_videos);
+        holder.bulkDownload = (IconView) convertView
+                .findViewById(R.id.bulk_download);
+        holder.bulkDownload.setIconColor(context.getResources().getColor(
+                R.color.edx_grayscale_neutral_base));
+        holder.numOfVideoAndDownloadArea = (LinearLayout) convertView
+                .findViewById(R.id.bulk_download_layout);
+        holder.rowSubtitlePanel =convertView.findViewById(R.id.row_subtitle_panel);
+        holder.halfSeparator = convertView.findViewById(R.id.row_half_separator);
+        holder.wholeSeparator = convertView.findViewById(R.id.row_whole_separator);
+
+        return holder;
+    }
+
+    public static class ViewHolder{
+        IconView rowType;
+        ETextView rowTitle;
+        ETextView rowSubtitle;
+        IconView rowSubtitleIcon;
+        IconView bulkDownload;
+        TextView noOfVideos;
+        LinearLayout numOfVideoAndDownloadArea;
+        View rowSubtitlePanel;
+        View halfSeparator;
+        View wholeSeparator;
+    }
+
+    public static class SectionRow {
+        public static final int ITEM = 0;
+        public static final int SECTION = 1;
+
+        public final int type;
+        public final boolean topComponent;
+        public final CourseComponent component;
+
+        //field to cache the temp value
+        public int numOfVideoNotDownloaded = -1;
+
+        public SectionRow(int type, CourseComponent component) {
+            this(type, false, component);
+        }
+
+        public SectionRow(int type, boolean topComponent, CourseComponent component) {
+            this.type = type;
+            this.topComponent = topComponent;
+            this.component = component;
+        }
     }
 }
